Webhook handling, Dockerfiles, deploy to production from CI (#49)
This PR adds an API route and handler for GitHub webhooks and adds new
jobs to the GitHub Actions workflow for building and deploying Docker
images.

### API Endpoint Updates:
*
[`api/src/endpoints/github.ts`](diffhunk://#diff-bf4c4fcf7a86713ef3bb8627fb8c82a4591b6bd3d1b067374b530aa49d24b371R48-R59):
Updated the `orgs` endpoint to use a constant for the GitHub API base
URL.
*
[`api/src/index.ts`](diffhunk://#diff-c7daeed68470fce500b64961dc1986f90d3c903d7679f553701de7aed7aa72a6L54-R83):
Added a new route for GitHub webhooks with a special raw body parser to
verify signatures and updated existing routes for better organization
and functionality.
[[1]](diffhunk://#diff-c7daeed68470fce500b64961dc1986f90d3c903d7679f553701de7aed7aa72a6L54-R83)
[[2]](diffhunk://#diff-c7daeed68470fce500b64961dc1986f90d3c903d7679f553701de7aed7aa72a6L75-R107)

### Webhook Handling:
*
[`api/src/endpoints/webhooks.ts`](diffhunk://#diff-28f0cd55d781fefa7c6c8e5ccba2de7eaf9acaec404a364bb5903ed6d4a101d2R1-R232):
Implemented a new function to verify GitHub webhook signatures, added a
handler for the `check_suite` webhook event, and included utility
functions for interacting with GitHub and the database.
*
[`api/src/endpoints/webhooks.test.ts`](diffhunk://#diff-39b083382f78b3e8ecc3e77f9fa612599d091107610a64530c03e4061e2702feR1-R90):
Added comprehensive tests for the `verifyWebhookSignature` function to
ensure correct verification of SHA-1 and SHA-256 signatures.

### CI/CD Pipeline Enhancements:
*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR83-R249):
Added new jobs for building and pushing Docker images for the backend
and worker, and deploying the frontend to S3 and the backend to EC2.

### Dependency Updates:
*
[`api/package.json`](diffhunk://#diff-d0ab9c1cc78fba971b5270314ad877706d340d099b42f46a406fc5de6a32fdf0L23-R23):
Updated the `octokit` dependency from version `4.1.1` to `4.1.2`.

### Miscellaneous Changes:
*
[`api/src/environment.ts`](diffhunk://#diff-b253a4e7a484867cf359d50cd6b31f9d8227611ac466050d20103d684e8dbc35L23-R23):
Simplified the `APP_URL` environment variable assignment.
*
[`api/src/index.test.ts`](diffhunk://#diff-331b00876e081406b54415b96065dc311b42867950e3f29b5007e18a18d370aaL15-R16):
Updated the root endpoint test to check the `/api` route instead of `/`.